requirements.txt
-i https://pypi.org/simple
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.61.0
flask==1.1.2
google-api-core[grpc]==1.22.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==1.20.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-core==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-datastore==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-ndb==1.5.0
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpcio==1.31.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.6.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
starlette==0.13.6; python_version >= '3.6'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1; python_full_version >= '3.6.1'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
